Move AttractionRecord class to its own file.
